requirements.txt
MDAnalysis
numpy
biopython